begining to use animations
changed the score saving method from core data to storing to a file. Still not able to store the score correctly. SetupViewController is still not loading the information from core data to display with the view appears. Dicovered that the API returns information different based on location. Need to add "if" statements to handle the discrepencies in order of the array/JSON data returned with the query