added question number retention for reviewing
Added the question number to the array for storage and linking. This will enable linking to the answers for studying only the questions that were answered wrong. Still need to change the data storage/controller so that the array is stored correctly in persistent storage and loaded from the controller as opposed to the model.